Refactorentrypoint (#68)
Update to the RenderPipes to support the new paradigm on how to go about writing a renderpipe.

Renderloops now should only contain configuration information + a call out to the rendering logic (preferably living in a static function). A loop now executes within a 'rendering context' that passes along a sidecar configuration file called a 'DataStore'. Any transient information that is needed between frames should be stored in this datastore (things like the materials / rendertextures). When the renderloop is destroyed this sidecar data is automatically cleaned up. It can also be cleaned manually.

Currently only the BasicRenderLoop has been ported to this new model due to the other loops not having a separation of concerns between transient data and configuration. They need the loop owners to detangle this before porting to the new model can take place. These existing loops still work, but they suffer from the same lifecycle issues they have had up until this point.

* [srl]Implement c# changes to match the c++ code changes.
* [srl]Fix scene view rendering of fptl and hd loop
* [srl]Start porting loops to use new state context.
* Add code to the loops to clear the intermediate renderers generated by the loop.
* Moving to interfaced based SRL for better abstraction from Unity core