Fix clear code in LightweightPipeline
The clear code wasn't clearing depth if the camera clear flag was set to Skybox.  The skybox rendering is dependent on depth being cleared initially, so it can fill in areas that opaque rendering hasn't covered.  But right now, depth is not being cleared, which breaks rendering.

This is a temporary fix for the clear code, as it isn't entirely optimal for TBDR GPUs.  Ideally, the first camera in the camera stack will clear both depth and color in order to prevent tile initialization.  Then the following cameras will avoid color clearing as they will need to load the contents from the previous camera render.  This code will be amended when Felipe adds camera stack support to LightweightPipeline.