Update forward only mechanism
Before we have ForwardOpaqueOnly and ForwardOpaqueDepthOnly and they
were required to be both present at the same time in sahder with regular
pass. Now pass are exclusive:

Guidelines: In deferred by default there is no opaque in forward.
However it is possible to force an opaque material to render in
forward
by using the pass "ForwardOnly". In this case the .shader should
not have "Forward" but only a "ForwardOnly" pass.
It must also have a
"DepthForwardOnly" and no "DepthOnly" pass as forward material (either
deferred or forward only rendering) have always a depth pass.