Remove OptimizeAffineMatrix()
Reason for removal: the function does not affect performance.
The last row of the matrix is only used to compute the W component of the vector resulting from the matrix-vector multiplication.
However, since we don't actually use the W component (we shouldn't...), the compiler doesn't use the last row during the multiplication.