Use RenderTextureDescriptor as argument, not HDCamera
Instead of passing around HDCamera, and fetching the RenderTextureDescriptor, we'll just pass the 'base' RenderTextureDescriptor, along with the fields we want to override.

The reason we pass in the list of overrides, instead of just modifying the descriptor is to keep a consistent appearance with the original API calls (e.g. GetTemporaryRT).  Additionally, we really want the base descriptor to contain the dimensions of the target, which will be consistent.  The changing parameters, such as depth buffer bits and color format, which should vary between RT creations.