Clear node selection when deleting nodes.
There was an issue that the selection list was not cleared on deletion of nodes.
This caused several problems when invoking node-editing actions right after
deletion.

e.g. Delete a random node and then open the right click menu without selecting
anything. This causes InvalidOperationException in BuildContextualMenu.

This small fix changes the behavior to clear the selection list in the end of
the deletion function.