Disable bindTextureMS for non-MSAA textures
For most pieces of RT state, if we need to create a non-MSAA RT, we can just copy over the state bits from the MSAA RT.  However, we shouldn't be turning on bindTextureMS on a 1x texture.  Unity somehow allows this creation, but then this leads to binding errors in the ShaderLab code.