Only initialize indirect buffer to size
On iOS, we get a GPU hang/reset based upon the execution of the ClearDispatchIndirect kernel. The buffer is created using the 'NUM_FEATURE_VARIANTS' constant (currently 27), but we're dispatching a threadgroup of 64 threads, so we are presumably going out of bounds (Metal GPU errors are not terribly descriptive, but this is a common cause).