Add light linking support (#1758)
* HDRP: Add support of light layering for forward (first draft)

TODO:
- Test instancing
- Check perf
- add an option to disable light layer support + in frameSettings
- add deferred rendering support (mean add an optional extra GBuffer)

* Add LightLinking support (version 1 - compile)

This PR add support of LightLayers feature.
To use it, setup mask renderingLayers on renderers and LightLayers on a lights. If a logical and of both give a non 0 value, the light is apply, otherwise it isn't.
This mean that by default LightLayers is 1 and RenderingLayers is 1. Then user can say LightLayers is only 2 and it will only affect renderingLayers with 2 or 3 mask value.

LightLayers require an extra GBuffer in deferred, mean it have extra bandwidth and memory cost.
LightLayers once enabled in HDRPAsset with SupportLightLayers can be control by FrameSettings. Mean a camera can enable extra RT only for an in game cinematic.

This PR refactor how the builtinData are send to LightLoop and other lighting function.
Before we were passing BakeLightingData everywhere. But it was containing only shadowMask and bakeDiffuseLighting
now we pass builtinData instead which embed shadowMask and bakeDiffuseLighting and all the others.
We add renderingLayers inside BuiltinData to be accessible everywhere.

This PR also refactor how we setup GBuffer. Now the deferredMaterial is fully in charge of allocating gbuffer and how to pack data inside these gbuffer.
This include ShadowMask and LightLayers buffer. Mean a lot of code have been cleanug, shadowmask is no more a separate buffer.

Add an extra GBuffer for storing RenderLayers for lightLayers feature.
Ambient occlusion is store in lightLayers buffer in deferred if available. This is used to avoid double occlusion when SSAO is enabled (SSAO will not darken AO but we take the min between SSAO and AO)

* Fix wrong encoding of rendering layers in deferred + FrameSettings UI

* Fix setup of GBuffer to always have the correct RT