Convert LW / Core RP tests into pacakges (#1829)
Port LW and Core tests to live in packages. They will be built as part of the package build process and also published. This means the can be used on katana and similar for testing.

LW / core are 'first version' of this. We need 'real usage' before we port HD / Shader Graph.