Screen Space Shadows Improvements (#2029)
* Don't render Screen Space Shadows if only directional light is available and it disables contact shadows.

When contact shadows is enabled in the Scene Settings, it is is still possible to disable the effect from the directional light itself. In that case, we don't need to run screen space shadows unless we have a dominant light.

* Use a single channel texture to hold screen space shadows

* Optimizing + improve quality of the contact shadow shader

Using a modified Interleaved Gradient Noise (from Next Generation Post Processing in Call of Duty: Advanced Warfare [Jimenez 2014]) that varies over time. If TAA is off, the IGN is not varying over time.

Some basic optimizations, early out and trying marching the depth buffer at half resolution until a hit is found, then switch back to full res to get the full intersection.
With this commit, on Scorpio on sample scene 1.18ms -> 0.815ms (0.36ms win)

* Add missing check for null m_CurrentSunLight

* Update changelog

* Break out of ray marching if we exit the sampling area

* Fix merge issues

* Improve vignette

It was assuming an arbitrary position of hit at mid ray if an occlusion was found. Replaced this with the actual position hit along the ray.

* Update and re-enable contact shadow screenshot tests

* Delete leftover meta files

* Remove leftover .meta file

* Use Load instead of sample to fetch depth