Use non-experimental UIElements (#2269)
### Purpose of this PR
Port the package to the non-experimental UIElements API.

Non-experimental UIElements API is in PR https://ono.unity3d.com/unity/unity/pull-request/76476/_/ui-elements/19.1-staging.

The plan is to make experimental and non-experimental APIs available together for a short while, before removing the experimental API (scheduled to happen before the 2019.1 cutoff).

This PR must land after the introduction of non-experimental API, but before the removal of experimental API.

---
### Release Notes
Please add any useful notes about the feature/fix that might be helpful for others.

---
### Testing status
**Katana Tests**: First off we need to make sure the Katana SRP tests are green?

**Manual Tests**: Open Shader graph examples in Unity. Checked the UI for obviously misplaced elements. Verified the shaders work.

**Automated Tests**: No new tests needed.

Any test projects to go with this to help reviewers? No

---
### Overall Product Risks
**Technical Risk**: Low to Medium. This is mainly trivial API changes. The riskiest changes are the input field refactor (every input is now inside a container field) and the default USS style adjustments which were done to match Unity style. These adjustments may impact UI element positioning.

**Halo Effect**: Low. Only UI.

---
### Comments to reviewers
Notes for the reviewers you have assigned.