HDRP: Disable 2004_AnimatedCookie due to side effect causing other test to fail
the test 2004_AnimatedCookie cause other test to fail with in this assert:
Debug.Assert(m_AutoSizedRTs.Count == 0, "RTHandle.Initialize should only be called once before allocating any Render Texture.");

Don't know why and it work on master. So certainly due to something on C++ side. When removing just the movie player it also work. so the issue is really related to playing movie and having other test using "something" that is already init.