Bug Fixes (#3334)
* Fixes for subgraph outputs issues

* Added safety check for naming outputs with numbers

There was a bug that if you started your output names with numbers it broke stuff

* Hiding the vars from the ui

* A lot of fixes

Removed unnecessary widgets from the inspector for subshader and shadergraph
Added active on Master nodes. Did this to dissallow deleting the active master nodes. Before this you could delete all master nodes.
This also made it so that we show only the active master node in the Main preview.
Also if you are trying to make a subgraph with nodes that are not allowed into subgraphs we prompt the user and let them know that those are not allowed and we can remove them from the selection and continue doing the subgraph.

* Fixes

Removed the dialog when converting to subgraph.
Fixed the main preview rendering only active Master node

* fixed the redrawing of the shadergraph window

* Fixed issues with the tests failing

need to catch the exception
and check if the node is not null. This is due to the test can create graphs with just nodes in without any master nodes.

* pr feedback fixes

* updated the changelog

* Update CHANGELOG.md