Get planar mapping working properly
This is probably a two-step fix: the planar itself probably just needed
some recalculations, but stuff went wrong elsewhere. We basically just
had bad luck that the planar was not "layer 0" as it is layer 0 that was
being resolved. I.e., there were two issues playing here. To be verified
what is still going wrong.