Add global validation for VT nodes
Checks if inter-node conditions are met. Particulary the fact that nodes cannot share texture input slots. (As a texture input slot can only belong to one particular stack). Similar two layers of the same not also cannot use the same input slot (since this would duplicate the data on both layers and since layers are sampled with the same UV's this is superfluous anyway.