Merge Hdrp/fix/case 1210058 to master (#5612)
* Fix for 1210058 - NaNs at certain view angles

Issue is caused by NaNs/INFs when computing basis on extremely thin (subpixel-width) geometry.  Putting some safeguards fixes the issue

* Update CHANGELOG.md to reflect 1210058 fix.

* Updated fix for issue #1210058

Overwrite the UVs to fix the derivatives.

* Updated fix for issue #1210058

Make sure the UVs are zero if unused to fix the derivatives.

* Undo last.change -- UV fix addresses NaNs

Zeroing the unused UV1 channel on the material seems to do the trick.

* Changelog adjustments to be moved

Current version puts everything in HDRP, so there is no change in core.

* Update CHANGELOG.md

Reflect fix for NaNs

* Fixed UV1 for lightmaps and detail maps.

* update comment

* Update NormalSurfaceGradient.hlsl

Co-authored-by: JulienIgnace-Unity <julien@unity3d.com>
Co-authored-by: sebastienlagarde <sebastien@unity3d.com>