Updated Submit for Alpha
Affected most of the visual changes from feedback. Now the template runs in a indirect baked mode with a mixed Directional Light that users can turn. The light is using distance shadowmasked so at close up it's using real-time shadows. Removed old assets that are no longer used and added the new bamboo setpiece. Additional setdressing in the form of decals/lights to the first room. Overall setup cleanup to present a cleaner setup of the scene setup and volumes as per feedback, and make sure there are no confusing settings. Also quickly tested that the scene can run with a simple camera controller.