Update HDRP Template for 2020.2 [Skip CI] (#868)
* First Commit for New HDRP template

First commit for the new HDRP template being worked on by the Asset Store Team

* Texture Update

Updated some of the textures to be of a smaller resolution, rebake lightmaps. WIP at the moment.

* WIP

Added new and tweaked textures. Cleaned up environment asset and renamed it accordingly. Remove other unused textures and materials.

* Commit

Added new Marble Texture set, and breaking up of main example asset environment so composition and placement of assets can be tested in engine easily.

* Texture Changes

* Update Marble_BC.png

* Commit

* Commit

* Commit

* Commit

Changes to Edge Decal Alpha. Did new woodplanks texture set to test.

* Asset Changes

-Added Woodplanks texture and material and apply it to first room.
-Clean up bonsai textures/material.
-Added a bit of metallic trim to steps to test ( will require a specific metal trim texture to make it nice.
-Adjustments to fog settings per room

* Commit

* Update EdgeDecals_BC.tga

* Commit for Alpha

Replace old files with current updated project. Added in the density volume feedback from today's meeting. Left with additional art tasks and lighting fixes to resolve

* Update template to latest version

* update manifest

* Updated Submit for Alpha

Affected most of the visual changes from feedback. Now the template runs in a indirect baked mode with a mixed Directional Light that users can turn. The light is using distance shadowmasked so at close up it's using real-time shadows. Removed old assets that are no longer used and added the new bamboo setpiece. Additional setdressing in the form of decals/lights to the first room. Overall setup cleanup to present a cleaner setup of the scene setup and volumes as per feedback, and make sure there are no confusing settings. Also quickly tested that the scene can run with a simple camera controller.

* Additional Lighting fixes

Adjusted the contact shadows of some of the volumes, rectified leaks with an additional shadowproxy mesh, and baking adjustments. Tweaked the smoothness of the wood material as well. Found another issue with the LODs of the small ceiling lights that needed adjustments.

* Update settings of HDRP assets

* Add FPS controlller

* Fix FPS controller prefabs, fix project version

* Add collider to the scene

* Disable shadow mask support + disable runtime debug to save compile time

* Update Readme.asset

* Remove preset

* Fix #extension in ffx_a.hlsl which cause warning

* fix collider issue (Trigger not setup correctly)

* Fix another trigger issue

* change screen space refraction to not have weight at border

* Add collider to avoid edge case with FPS controler. Move start position

* Update HDRenderPipelineAsset.asset

* Revert "Fix #extension in ffx_a.hlsl which cause warning"

This reverts commit 4612d20c87bb3039238be5c46c1d160b7073b5db.

Co-authored-by: AdrianQuek <adrian.quek@unity3d.com>