Update eye index processing for Vulkan
Due to an ancient issue with the ARM Mali shader compiler, a special hack needed to be done to get GLES to work properly.  This workaround isn't required for Vulkan, so we undo most of the hack.  The global constant buffer is still defined (however, it's declared empty for Vulkan) to minimize changes to the engine.