XRE-1791: SRP will set the stereo render mode it supports.
For URP if multiview is supported on the build target it will let the shader compiler to know to add that variant. If not it will use SPI.

For HDRP it will check if the build target supports SPI and if it does it will tell the shader compiler to add that varaint.