Texture Cleanup
Resize some of the textures and make sure they are saved with the highest compression settings for PNG. Remove unused rock textures that slipped the previous cleanup. Also resave out a 2K albedo for the Rock_A, with its mask at 512 and normal at 1024. Current size of template is within 100mb.