Temp. DXC compilation fix for SSR and SSS shaders (forces DXC use)
*Add new 'wavebasic' requirement pragmas to SSR and SSS, this forces DXC on
and sets an appropriate MSL minspec for the platform ('#pragma use_dxc' is
insufficient as its minspec is lower than wave ops. on Metal devices).
These shaders worked without DXC as the intrinsics they use were emulated
with an overloaded function. This would break with DXC as DXC gives the
acutal intrinsics preference when parsing and we end up running SPIRV-Cross
(MSL gen) win an insufficient MSL version. Temporary commit, should not
be allowed to ship to SRP.  # Please enter the commit message for your
changes. Lines starting # with '#' will be ignored, and an empty message
aborts the commit.