Added feature: User can specify a directory that holds their VFX template assets (#34)
Added feature: User can specify a directory that holds their VFX template assets

Allow VFX artists to specify an asset directory that holds their own VFX templates aka Systems. This expands on the Systems functionality seen in the Node menu.

## How to use
The user directory is set in the menu: Edit -> Preferences -> Visual Effects -> User Systems
![image (1)](https://github.cds.internal.unity3d.com/storage/user/2318/files/d738c680-a1bb-11ea-94d4-68711743821d)

Open a VFX asset in the VFX Graph. Press spacebar to open the Node menu and note the new User Systems option.
![image (2)](https://github.cds.internal.unity3d.com/storage/user/2318/files/20891600-a1bc-11ea-8ab4-831989aef9df)

This option points to the User Systems directory specified by the user. This option will only show up if the user has set the directory to a valid location. It supports nested folders.
![image (3)](https://github.cds.internal.unity3d.com/storage/user/2318/files/4d3d2d80-a1bc-11ea-954f-4c29da91253f)

Compare this to the folder structure in Assets. Note that empty folders and folders not containing VFX assets are not displayed:
![image (4)](https://github.cds.internal.unity3d.com/storage/user/2318/files/7cec3580-a1bc-11ea-86e1-fa3154e108a2)