Fix warning "implicit truncation of vector type"
When creating a new URP Lit shader graph, changing the input color in any way results in an HLSL warning, "implicit truncation of vector type", in the generated `BuildVertexDescriptionInputs()` function, because `output.ObjectSpaceTangent` is of type `float3` and `input.tangentOS` is of type `float4`.  To avoid the warning, an `.xyz` swizzle is used to explicitly convert to `float3`.