Shadergraph/documentation/sg-doc-2627 (#3620)
* Creation of new View Vector Node doc

* Update View Direction node. 

The PR containing the code changes for View Direction / View Vector also included a change to this article. The developer and editorial had reviewed and approved the documentation prepared to release with the code PR, but I was waiting to hear what version I should commit the change to. Due to a miscommunication, I did not have the opportunity to modify this article before the push of the View Direction / View Vector code. Now I am doing so. 
The text I have replaced is:
"Provides access to the mesh vertex or fragment's **View Direction** vector. This is the vector from the vertex or fragment to the camera. The coordinate space of the output value can be selected with the **Space** dropdown parameter.

NOTE: In versions prior to 11.0, the **View Direction** vector was not normalized in the **Universal Render Pipeline**. Version 11.0 changed this behavior, and this vector is now normalized in both the **High-Definition Render Pipeline** and the **Universal Render Pipeline**. To mimic old behavior, you can use the [Position Node](Position-Node.md) in **World** space and subtract the **Position** output of the [Camera Node](Camera-Node.md)."

In the version of the text approved by the dev and editorial, the solution to mimic old behavior is to use the View Vector mode. So I have replaced mention of the position and camera nodes with mention of the View Vector node.

* changelogged DOC2627 changes

* Reformat files

Co-authored-by: Théo Penavaire <theo.penavaire@unity3d.com>