Added descriptions to labels. Got saving to work.
Added the object path as well as the file path for the documents. Sort the documents based on ID so it won't continue to open different scene files.
Implement saving of the scenes and hack in a way to save only once (without saving the initial scene). Worst case this is an example of how to do this for a single converter, but ideally if someone is using search, we should probably provide methodology for sorting each converter together.
Updated the default material remapping to include every type. Still need to create some materials for some of the types.
Remove default materials folder since we can mostly just replace with the existing URP materials directly.