Fixed a bug that was crashing in indexing and other areas. Fixed replacement of multimaterials.
Occasionally TextMeshPro is missing GetMethods for Shared Font materials. I have no idea why this is happening, but it's happening consistently enough that we should handle it. Now it throws a warning in that case, since it can't do anything about it.
Rearranged and commented the public methods in the static class. I'm realizing we might want to make these internal instead of public.
Added the case to support multi-material reassignment as well.