Fixes from the PR as well as updates to accommodate the new OnRun method.
This includes all of the fix suggestions from the PR.
This also include the updates to loading objects resulting from the change in the OnRun method. Now that that method contains a single object rather than the list of objects, the loading code handles them one at a time. This actually simplifies the loading code in a good way.