Removed forced clear on mobile
Previously, we would always clear each camera target on mobile, even when the camera's target's background was set to "Don't care". This was based on the assumption that a clear comes for free on mobile. However, this assumption is not always true.
- Not all mobile GPUs are tile-based (even though the vast majority are, Tegra being an exception), so the assumption that clearing is always done through a load action on mobile is incorrect
- An empty render pass with a DontCare load action will be optimized away low level code. However, an empty render pass with a Clear load action cannot be optimized away, because we would expect the render target to be cleared afterwards. This means that, by forcing a clear on mobile, we sometimes take away an opportunity for low level code to optimize away a renderpass. This currently happens with the 2DRP

Based on these downsides, I removed the forced clear for mobile when setting a camera's background to "Don't care"