Guarantee that the scene gets saved if a change has been made
In the initial implementation of this, there was some trickery to try not to save the scene on every change, but with the change to not running on a list of objects in the context that resulted in edge cases where the scene would not be properly saved with the changes.
This guarantees that changes get saved. Unfortunately, it does resave the scene every time a change is made. There is a way around this, but it would require the framework to be more aware of when it changed from one converter to the next and save the current scene in between converters. I think for now this isolates the converters and doesn't require the framework to make assumptions about whether a scene should be saved or not. It will be a little slower to run, but with the Indexing taking so long, this part of it will end up being fairly trivial.