Update reference image for 049 after OIDN upgrade
Case 1353374: Gfx test UniversalGraphicsTests.049_Lighting_Mixed_Subtractive is failing. The scene is set to 'auto lightmap filtering' which means that Optix denoiser is used when available and we fall back to OIDN if not.
On Bokken, the driver is quite old so the Optix denoiser isn't used. When manually setting filtering to OIDN I can reproduce the artifacts locally. Therefore I am accepting the new output as reference image.