Added additional synonyms to these nodes for improved searchability
Added synonyms for these nodes so that when you do a search for them in the Create Node menu, you'll get more results from a broader range of search terms.  This should make it easier for people to find the nodes their looking for even if they don't know the exact name.