Fixed Bug That Prevented Sclera Iris Blend node from being added
The Sclera Iris Blend node had the same name as the Sclera Limbal Ring node, so when you selected the Sclera Iris Blend node from the Create Node menu, the Sclera Limbal Ring node was added instead.  This made it impossible to add a Sclera Iris Blend node to the graph.