Add support for single pass multiview
the stereo rendering mode single pass multiview was simply not handled as a case. It is meant to be quite similar to the single pass instance that renders to the texture array so this change just adds the single pass multivew check wherever we were checking for SPI and this seems to fix the problem of the missing eye.