Common FSR Utilities for SRPs
This change adds two new utility files to SRP core to make it easier
to use FSR in SRPs.

The FSRUtils C# class contains a port of the FSR constant values setup
code from the HLSL files. This is needed for SRPs that do not support
writing to buffers from inside a shader.

The FSRCommon HLSL include contains helper functions for using EASU
and RCAS inside the Unity shader environment.

The two files must be used together since they are tightly coupled by
the identifiers used for the shader constants.