StackLit coat ssr-rtr fixes:
Add support for more refined reflection hierarchy for SSR-RTR blending with IBL and to avoid double coat lighting.
This is mainly based on https://github.com/Unity-Technologies/Graphics/pull/4968/, but with the added feature to re-use the coat-traced light for similarly-rough bottom layer lobes, like done for Lit.

This fixes several issues in the implementation in StackLit since #5565:

-Similar issues fixed for Lit (see other PR for Lit)
-StackLit can have a rough coat, but an arbitrary 0.9 smoothness value was used to decide if the coat-traced light would be re-used for the bottom lobes.
-The wrong bottom lobe roughnesses were used (values that don't include the effect of BRDF vertical layering computations)
-StackLit can have different lobe directions and the light hierarchy now account for this.