Move FXAA Logic to Common.hlsl
This change moves the FXAA HLSL implementation into the common post
processing shader file. This will make it easier to execute the FXAA
shader logic outside of the FinalPost pass which is a prerequisite for
FSR.

This change also modifies the FXAA Load helper function to make it use
point sampling instead of linear sampling on GLES. This should yield
more consistent behavior between GLES and non-GLES environments.