Disable 16-Bit FSR Implementation
This change prevents URP from using the 16-bit FSR implementation since
it seems to have compatibility issues on some platforms which result in
a black screen. This has been observed on the Nvidia DX11 driver, and
possibly some XR platforms as well.

Using the 32-bit implementation will negatively impact performance on
mobile platforms, so this change should be removed as soon as the
problems are resolved.