Add FSR Sharpness Control to HDRP
This change adds a new GUI control in HDRP's camera properties that
allows users to control the intensity of the sharpness pass used by
FSR. The sharpness intensity is specified as a value from 0.0 (least
sharp) to 1.0 (most sharp).

Adding a 0.0 -> 1.0 interface for sharpness required some special
mapping logic since FSR expects a non-linear sharpness value. This
logic was added to SRP Core as a small wrapper function called
"SetRcasConstantsLinear".

When the sharpness parameter is set to zero, the post processing logic
now disables the RCAS shader code to improve performance since it won't
make a visible difference.