Fix Scene View + Render Scale Logic
This commit reverses an unintentional change to the scene view's
handling of the render scale setting. This change was introduced
in #6274.

This commit also attempts to consolidate some of the render scale logic
in URP's code by modifying the shader constant logic to query the scaled
camera dimensions directly from the camera target descriptor rather than
calculating them again.