Removed Approximation Function Usage in 16-bit FSR
Usage of shader function approximations was causing a black screen on
Nvidia DX11 drivers. Even the higher precision variants also led to
visual artifacts so the approximations have been replaced completely in
the 16-bit path.

Strangely enough, these functions do not cause issues on the DX12 and
Vulkan paths, so there may be some shader compilation difference that we
need to address instead of changing this.