Expose FSR Sharpness to GUI Controls
This change adds a new property to URP's pipeline asset that allows
users to specify how much sharpening they want the FSR upscaling filter
to perform. This value is specified in a 0.0 to 1.0 linear range with
0.0 being the least sharp and 1.0 being the most sharp.

This property is also gated behind a checkbox which allows us to
identify when the user wants the default sharpness value or a custom
value.