Re-Enable 16-bit FSR
This change re-enables the 16-bit FSR implementation but adds some extra
logic to keep it off on DX11 since that's the only known place where the
issues occur.

Note: The issues seem to be relating to the use of the 16-bit math
approximation functions in both EASU and RCAS. When those are replaced
with standard functions, the shaders work normally. Rather than
modifying the shaders to stop using the approximations, it makes more
sense for us to just turn 16-bit mode off for DX11 since it wasn't
likely to benefit from it anyways.