[HDRP] Add FSR Sharpness Control (#6373)
* Add FSR Sharpness Control to HDRP

This change adds a new GUI control in HDRP's camera properties that
allows users to control the intensity of the sharpness pass used by
FSR. The sharpness intensity is specified as a value from 0.0 (least
sharp) to 1.0 (most sharp).

Adding a 0.0 -> 1.0 interface for sharpness required some special
mapping logic since FSR expects a non-linear sharpness value. This
logic was added to SRP Core as a small wrapper function called
"SetRcasConstantsLinear".

When the sharpness parameter is set to zero, the post processing logic
now disables the RCAS shader code to improve performance since it won't
make a visible difference.

* Reworked HDRP FSR Sharpness Override

This change takes the existing per-camera sharpness setting, and
replaces it with a paired pipeline asset + per camera override model
similar to the one used by DLSS. This change also adds extra logic to
ensure that FSR specific settings are hidden when FSR is not currently
being used.

* Add Pipeline Level FSR Sharpness Override Switch

This change adds a top level "override enabled" checkbox to the render
pipeline asset which now gates the FSR sharpness setting.

This approach has the following benefits:
- Allow us to easily identify the user's intention (default vs custom)
- Makes it easy for users to restore "default" behavior
- Makes it easy for us to change the "default" value in the future
  without requiring changes to existing projects

* Moved Back to Slider Based UI Control for Pipeline

This change modifies the pipeline level FSR sharpness control to use a
slider again instead of a delayed float input control. Once the issue
with changing pipelines settings is fixed, this works just as well as
the delayed float control.

* Updated Changelog

This change adds a line to the changelog for the FSR sharpness
override.

* Updated SRP Core Changelog

* Documentation For FSR Sharpness Controls

This commit adds some documentation for the new FSR sharpness controls
to the dynamic resolution markdown file.

* Fixed Formatting